src/utils/config.py
"""Configuration loading utilities for the sentiment dashboard."""

import logging
from dataclasses import dataclass, field
from pathlib import Path
from typing import Any

import yaml

logger = logging.getLogger(__name__)


@dataclass
class Config:
    """Application configuration loaded from YAML."""

    database: dict[str, Any] = field(default_factory=dict)
    model: dict[str, Any] = field(default_factory=dict)
    training: dict[str, Any] = field(default_factory=dict)
    dashboard: dict[str, Any] = field(default_factory=dict)
    api: dict[str, Any] = field(default_factory=dict)
    simulator: dict[str, Any] = field(default_factory=dict)
    preprocessing: dict[str, Any] = field(default_factory=dict)
    topic_modeling: dict[str, Any] = field(default_factory=dict)
    trend_detection: dict[str, Any] = field(default_factory=dict)


def load_config(path: str = "configs/config.yaml") -> Config:
    """Load application configuration from a YAML file.

    Args:
        path: Path to the YAML configuration file.

    Returns:
        Config dataclass populated with the YAML data.

    Raises:
        FileNotFoundError: If the config file does not exist.
        yaml.YAMLError: If the YAML is malformed.
    """
    config_path = Path(path)
    if not config_path.exists():
        raise FileNotFoundError(f"Configuration file not found: {path}")

    with open(config_path) as f:
        data = yaml.safe_load(f)

    if not isinstance(data, dict):
        raise ValueError(f"Invalid configuration format in {path}")

    logger.info("Loaded configuration from %s", path)
    return Config(**{k: v for k, v in data.items() if k in Config.__dataclass_fields__})
